feat: migrate user tests to work with forum v2
- Migrated tests from rest_api/test_api, rest_api/test_views, and base/tests.py
