+++ lms/djangoapps/discussion/rest_api/tests/utils.py
@@ -675,7 +675,7 @@ class ForumMockUtilsMixin(MockForumApiMixin):
         }
         self.set_mock_side_effect('get_user', make_user_callbacks(self.users_map))
 
-    def register_get_user_retire_response(self, user, body=""):
+    def register_get_user_retire_response(self, user, status=200, body=""):
         self.set_mock_return_value('retire_user', body)
 
     def register_get_username_replacement_response(self, user, status=200, body=""):
