+++ openedx/core/djangoapps/django_comment_common/comment_client/course.py
@@ -8,7 +8,6 @@ from edx_django_utils.monitoring import function_trace
 from opaque_keys.edx.keys import CourseKey
 
 from forum import api as forum_api
-from openedx.core.djangoapps.django_comment_common.comment_client import settings
 
 
 def get_course_commentable_counts(course_key: CourseKey) -> Dict[str, Dict[str, int]]:
